+++ opencti-platform/opencti-front/src/private/components/analysis/external_references/CyioCoreObjectExternalReferences.js
@@ -6,14 +6,12 @@ import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
-import Avatar from '@material-ui/core/Avatar';
 import Skeleton from '@material-ui/lab/Skeleton';
 import { QueryRenderer as QR } from 'react-relay';
 import DarkLightEnvironment from '../../../../relay/environmentDarkLight';
 import inject18n from '../../../../components/i18n';
-import { QueryRenderer } from '../../../../relay/environment';
+// import { QueryRenderer } from '../../../../relay/environment';
 import CyioCoreObjectExternalReferencesLines, {
   cyioCoreObjectExternalReferencesLinesQuery,
 } from './CyioCoreObjectExternalReferencesLines';
@@ -111,69 +109,6 @@ class CyioCoreObjectExternalReferences extends Component {
             );
           }}
         />
-        {/* <QueryRenderer
-        query={cyioCoreObjectExternalReferencesLinesQuery}
-        variables={{ id: cyioCoreObjectId, count: 200 }}
-        render={({ props }) => {
-          if (props) {
-            return (
-              <CyioCoreObjectExternalReferencesLines
-                cyioCoreObjectId={cyioCoreObjectId}
-                data={props}
-              />
-            );
-          }
-          return (
-            <div style={{ height: '100%' }}>
-              <Typography
-                variant="h4"
-                gutterBottom={true}
-                style={{ float: 'left', marginBottom: 15 }}
-              >
-                {t('External references')}
-              </Typography>
-              <div className="clearfix" />
-              <Paper classes={{ root: classes.paper }} elevation={2}>
-                <List>
-                  {Array.from(Array(5), (e, i) => (
-                    <ListItem
-                      key={i}
-                      dense={true}
-                      divider={true}
-                      button={false}
-                    >
-                      <ListItemIcon>
-                        <Avatar classes={{ root: classes.avatarDisabled }}>
-                          {i}
-                        </Avatar>
-                      </ListItemIcon>
-                      <ListItemText
-                        primary={
-                          <Skeleton
-                            animation="wave"
-                            variant="rect"
-                            width="90%"
-                            height={15}
-                            style={{ marginBottom: 10 }}
-                          />
-                        }
-                        secondary={
-                          <Skeleton
-                            animation="wave"
-                            variant="rect"
-                            width="90%"
-                            height={15}
-                          />
-                        }
-                      />
-                    </ListItem>
-                  ))}
-                </List>
-              </Paper>
-            </div>
-          );
-        }}
-      /> */}
       </>
     );
   }
