+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/CyioCoreObjectHistoryLines.js
@@ -20,7 +20,7 @@ const styles = () => ({
 class CyioCoreObjectHistoryLinesComponent extends Component {
   render() {
     const {
-      t, classes, data, isRelationLog,
+      classes, data, isRelationLog,
     } = this.props;
     const logs = pathOr([], ['logs', 'edges'], data);
     return (
